pyproject.toml
[tool.poetry]
name = "sample"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.23"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
jinja2 = "^3.1.2"
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"
reportlab = "^4.0.7"
python-pptx = "^0.6.23"
lxml = "^4.9.3"
aiofiles = "^23.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
